Re-implement VisualManager Add new visual effects for: >Titan >Revenant >Hornet >Healbot >Brawler >Tower Upgrades and Stunned New Attack Type: >Ricochet >Split >Orbit
.//.Tomorrow's Tasks:
Attack texturing and polishing
Wave configuration
Tower stat changes

im taking too long, sorry :<

+++ src/Game/Attack/Projectile.h
@@ -8,7 +8,7 @@ private:
 	Vector2 AttackProjectileDirection;
 	Color AttackColor;
 public:
-	static const Attack* ProjectileTemplateBuildAndGet(const Vector2& _AttackStartPositon, const Vector2& _AttackDestinationPosition, const float& _AttackMovementSpeed, const float& _AttackDamage, const int& _AttackTargetID, const int& _AttackOwnTowerID, const size_t& _Lifespan, const Color& _AttackColor);
+	static const Attack* ProjectileTemplateBuildAndGet(const Vector2& _AttackStartPositon, const Vector2& _AttackDestinationPosition, const float& _AttackMovementSpeed, const float& _AttackDamage, const int& _AttackTargetID, const int& _AttackOwnTowerID, const int& _Lifespan, const Color& _AttackColor);
 	void Draw() const override;
 	void Update() override;
 };
